fix type-hint

+++ src/aoc/2024/13.py
@@ -23,7 +23,7 @@ data = aoc.get_data_from_file(13, 2024)
 
 A_COST: int = 3
 B_COST: int = 1
-CONV_ERROR: int = 1e13
+CONV_ERROR: float = 1e13
 
 
 def parse_data(data: str):
